All: Major refactors in callback handling
Now we use signals (mostly static ones in the singletons) to move the
update logic from the signal generators to the ones who actually update
their status because of the signal.

Also added support for adding elements into the Graph, although I feel
there's too much Glib::ustring copies going on, and the overall logic
could possibly be way less complex... But for now it seems to work! :)