IO: Move dialog out of gstreamer/ and linearize it
Instead of using signals to tell the involved parties that a pipeline
was added, now we call them directly, since previously it was confusing
who called the signal and who reacted to it. Now it's clearer who is
doing what.
The biggest problem with this implementation is that PipelineFactory is
checking for the Client type to decide what to do. Ideally we would have
the Client tell it what to do via a virtual function call that GtkClient
would override. In fact, a framework where the Client would tell the
application parts who is responsible for what task in general seems like
a good idea for now. We'll see :p