Graph: Remove general 'Property' in favor of a local one in Element
The previous implementation of 'Property' was awkward in that it was
confusing what was a field, what was a property, and what was a
propertylist, as well as transforming every entry in a std::string.

The new implementation is local to Element, and use variadics to store
the values of properties. It is still necessary to go through the
available types to determine what to do in that case, but this logic is
now entirely local to Property, easing its maintanence.

The biggest problem with the way PipelineEditor is currently implemented
is that it needs to know about what type of Node it is constructing a
properties widget for. As with the previous commit, the idea to make this
better will be to move responsability of creating the widget to someone
else, in this case to the ElementNode.